+++ Holovibes/includes/tools/info_manager.hh
@@ -61,7 +61,7 @@ namespace holovibes
 		/*! Stop to refresh the info_panel display*/
 		void stop_display();
 
-		void insertInputSource(const int width, const int height, const int depth);
+		void insertInputSource(const int width, const int height, const size_t depth);
 
 		/*! Add your information until is remove, and call draw()
 		** \param key is where you can access to your information
